requirements.txt.txt
mysql-connector    2.2.9
pip                23.2.1
